# Lazy load objects initialization