Continuation of #43

Instead of caching at build time, I just lazy load the initialization of the objects:

Before:

```bash
$ hyperfine --warmup 3 "node ./index.js" 
Benchmark 1: node ./index.js
  Time (mean ± σ):     109.1 ms ±   1.8 ms    [User: 96.6 ms, System: 36.7 ms]
  Range (min … max):   102.8 ms … 112.9 ms    27 runs

```

After:

```bash
$ hyperfine --warmup 3 "node ./index.js"
Benchmark 1: node ./index.js
  Time (mean ± σ):     105.2 ms ±   1.1 ms    [User: 93.4 ms, System: 34.8 ms]
  Range (min … max):   103.4 ms … 108.7 ms    27 runs

```