@H4ad 2024-04-06
> Do you license your work here under [Apache 2.0](https://apache.org/licenses/LICENSE-2.0)?

Yes, no problem.