@kemitchell 2024-04-11
Do I correctly understand that the benefit here, at least on the version of Node tested, is on the order of many a single millisecond during initial load? That in exchange for calling a function to load each variable?

I think it's safe to assume that where this module gets loaded, it will also be invoked. If it's not being invoked, the perf fix would be to not load in the first place.