@kemitchell 2024-04-11
Would you expect a similar 4ms delay on first invocation?