@kemitchell 2024-04-11
Will the first call to the function the package exports take an additional 4ms to initialize all the variables now returned from the loader helper functions?