@H4ad 2024-04-11
> Will the first call to the function the package exports take an additional 4ms to initialize all the variables now returned from the loader helper functions?

Yes.

The overall time will be the same, we save 4ms during load time but we spent 4ms in the first call for the package.