@kemitchell 2024-04-12
I don't mean to see overly resistant here. I appreciate you taking the time to answer my questions!

If the trade-off is just _when_ initial loading occurs, I wonder why it's better for the package to prescribe data loading on invocation over preemptive data loading on module load. Particularly since `require` and dynamic `import()` allow users to defer package loading without any changes to the implementation of the package, but there's no existing mechanism for making a package coded to load on invocation to load preemptively, instead.

This package is in many ways a bunch of plain-text lists and mappings, pieced together with relatively thin layers of glue code. So not the worst place to look for a load-related perf win or load-time/runtime arbitrage. And `npm run` may indeed have been suffering, with me still wondering why it loads this package at all. But I'm not sure fairly rote wrappings of global vars in getter functions really makes sense.