@H4ad 2024-04-12
> Particularly since require and dynamic import() allow users to defer package loading without any changes to the implementation of the package

Yes, but people usually don't know how much time each dependency takes to load, I'm working on Node to add a feature to let them know the amount of time that takes to load but people usually think is zero-cost or something really fast (1~3ms).

> And npm run may indeed have been suffering, with me still wondering why it loads this package at all.

I removed this package from the hot-path, now this package is only loaded when NPM needs to validate licenses (when run npm init, for example).

If you think that is not worth introducing this change, that's ok for me since my main issue was to reduce load time of `npm run` only.