@kemitchell 2024-04-15
@H4ad I just added a performance note about load time and dynamic `import()` to README:

0bf80b1c488d42759ead43a23b16741b782433b6

If hope you don't mind that I thanked you by name! Thanks again for your issue and PR here.